Refactor group details screen UI
This commit refactors the group details screen for a cleaner user interface.

- Group information (name, day, grade) is now displayed in a dedicated "Group Info" section with an edit icon.
- Students are listed in a separate "Students" section.
- Added a "Start session" button.
- Introduced `GroupStudentItemWidget` for displaying student details within the group.
- Updated theme colors for `scaffoldBackgroundColor` and `cardColor`.
- Made `onDeleteClick` optional in `StudentItemWidget`.

+++ lib/widgets/students/student_item_widget.dart
@@ -3,7 +3,6 @@ import 'package:teacher_app/screens/students_list/states/student_item_ui_state.d
 import 'package:teacher_app/widgets/delete_icon_widget.dart';
 import 'package:teacher_app/widgets/grade_with_icon_widget.dart';
 import 'package:teacher_app/widgets/phone_with_icon_widget.dart';
-
 import '../../themes/txt_styles.dart';
 import '../app_txt_widget.dart';
 import '../forward_arrow_widget.dart';
@@ -12,13 +11,13 @@ class StudentItemWidget extends StatelessWidget {
 
   final StudentItemUiState uiState;
   final Function(StudentItemUiState) onItemClick;
-  final Function(StudentItemUiState) onDeleteClick;
+  final Function(StudentItemUiState)? onDeleteClick;
 
   const StudentItemWidget(
       {super.key,
       required this.uiState,
       required this.onItemClick,
-      required this.onDeleteClick});
+        this.onDeleteClick});
 
   @override
   Widget build(BuildContext context) {
@@ -40,6 +39,7 @@ class StudentItemWidget extends StatelessWidget {
                 children: [
                   _studentName(),
                   Spacer(),
+                  if(onDeleteClick != null)
                   _deleteIcon(),
                 ],
               ),
@@ -69,5 +69,5 @@ class StudentItemWidget extends StatelessWidget {
 
   _studentName() => AppTextWidget(uiState.name, style: AppTextStyle.title);
 
-  _deleteIcon() => DeleteIconWidget(onClick: (){onDeleteClick(uiState);});
+  _deleteIcon() => DeleteIconWidget(onClick: (){onDeleteClick?.call(uiState);});
 }
